refactor and make flatter heirarchy
