+++ nbox/__init__.py
@@ -1,8 +1,10 @@
 from .model import Model
 from .load import load, PRETRAINED_MODELS
-from .jobs import Instance, Job
+from .jobs import Job
+from .instance import Instance
 from .auth import AWSClient, GCPClient, OCIClient, DOClient, AzureClient
 from .operators import Operator
 from .utils import logger
+from .subway import Sub30
 
 __version__ = "0.8.7"
