--- nbox/auto/__init__.py
@@ -1,6 +1,0 @@
-"""
-Auto (WIP)
-==========
-
-Auto will do what auto should do.
-"""
